cleanup in configs
