+++ cmd/captplanet/main.go
@@ -18,24 +18,8 @@ import (
 
 	"storj.io/storj/internal/fpath"
 	"storj.io/storj/internal/memory"
-	"storj.io/storj/pkg/accounting/rollup"
-	"storj.io/storj/pkg/accounting/tally"
-	"storj.io/storj/pkg/audit"
-	"storj.io/storj/pkg/bwagreement"
 	"storj.io/storj/pkg/cfgstruct"
-	"storj.io/storj/pkg/datarepair/checker"
-	"storj.io/storj/pkg/datarepair/repairer"
-	"storj.io/storj/pkg/discovery"
-	"storj.io/storj/pkg/inspector"
-	"storj.io/storj/pkg/kademlia"
-	"storj.io/storj/pkg/miniogw"
-	"storj.io/storj/pkg/overlay"
-	"storj.io/storj/pkg/piecestore/psserver"
-	"storj.io/storj/pkg/pointerdb"
 	"storj.io/storj/pkg/process"
-	"storj.io/storj/pkg/provider"
-	"storj.io/storj/pkg/satellite/satelliteweb"
-	"storj.io/storj/pkg/server"
 )
 
 var (
@@ -49,51 +33,6 @@ var (
 	defaultConfDir = fpath.ApplicationDir("storj", "capt")
 	confDir        *string
 )
-
-// Config defines broad Captain Planet configuration
-type Config struct {
-	SatelliteCA         provider.CASetupConfig
-	SatelliteIdentity   provider.IdentitySetupConfig
-	UplinkCA            provider.CASetupConfig
-	UplinkIdentity      provider.IdentitySetupConfig
-	StorageNodeCA       provider.CASetupConfig
-	StorageNodeIdentity provider.IdentitySetupConfig
-	ListenHost          string `help:"the host for providers to listen on" default:"127.0.0.1"`
-	StartingPort        int    `help:"all providers will listen on ports consecutively starting with this one" default:"7777"`
-	APIKey              string `default:"abc123" help:"the api key to use for the satellite"`
-	EncKey              string `default:"insecure-default-encryption-key" help:"your root encryption key"`
-	Overwrite           bool   `help:"whether to overwrite pre-existing configuration files" default:"false"`
-	GenerateMinioCerts  bool   `default:"false" help:"generate sample TLS certs for Minio GW"`
-
-	Satellite    Satellite
-	StorageNodes [storagenodeCount]StorageNode
-	Uplink       miniogw.Config
-}
-
-// Satellite is for configuring client
-type Satellite struct {
-	Server      server.Config
-	Kademlia    kademlia.SatelliteConfig
-	PointerDB   pointerdb.Config
-	Overlay     overlay.Config
-	Inspector   inspector.Config
-	Checker     checker.Config
-	Repairer    repairer.Config
-	Audit       audit.Config
-	BwAgreement bwagreement.Config
-	Web         satelliteweb.Config
-	Database    string `help:"satellite database connection string" default:"sqlite3://$CONFDIR/master.db"`
-	Discovery   discovery.Config
-	Tally       tally.Config
-	Rollup      rollup.Config
-}
-
-// StorageNode is for configuring storage nodes
-type StorageNode struct {
-	Server   server.Config
-	Kademlia kademlia.StorageNodeConfig
-	Storage  psserver.Config
-}
 
 func main() {
 	go dumpHandler()
