+++ cmd/storagenode/main.go
@@ -28,6 +28,22 @@ import (
 	"storj.io/storj/pkg/storj"
 )
 
+// StorageNode defines storage node runtime configuration
+type StorageNode struct {
+	Server   server.Config
+	Kademlia kademlia.StorageNodeConfig
+	Storage  psserver.Config
+}
+
+// SetupStorageNode defines storage node setup configuration
+type SetupStorageNode struct {
+	CA        identity.CASetupConfig
+	Identity  identity.SetupConfig
+	Overwrite bool `default:"false" help:"whether to overwrite pre-existing configuration files"`
+
+	StorageNode
+}
+
 var (
 	rootCmd = &cobra.Command{
 		Use:   "storagenode",
@@ -56,16 +72,9 @@ var (
 		RunE:  cmdDiag,
 	}
 
-	runCfg struct {
-		Server   server.Config
-		Kademlia kademlia.StorageNodeConfig
-		Storage  psserver.Config
-	}
-	setupCfg struct {
-		CA        identity.CASetupConfig
-		Identity  identity.SetupConfig
-		Overwrite bool `default:"false" help:"whether to overwrite pre-existing configuration files"`
-	}
+	runCfg   StorageNode
+	setupCfg SetupStorageNode
+
 	diagCfg struct {
 	}
